# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/projects": {
      "get": {
        "operationId": "ProjectsController_list",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "List projects in a workspace",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "ProjectsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Create a project",
        "tags": [
          "projects"
        ]
      }
    },
    "/v1/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Project detail",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "ProjectsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Rename a project / edit its description",
        "tags": [
          "projects"
        ]
      },
      "delete": {
        "operationId": "ProjectsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "4XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Error envelope — all non-2xx responses conform to ApiErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Delete an empty project (409 when it still holds documents)",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Knowledge API",
    "description": "Dynamic Knowledge Platform — documents, revisions, ingestion, search",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "kn_ API key or ks_ session token (AUTH_MODE=api-key)"
      }
    },
    "schemas": {
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "workspaceId",
          "name"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404,
            "description": "HTTP status (0 is reserved for client-synthesized errors)"
          },
          "code": {
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "VALIDATION_FAILED",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "PAYLOAD_TOO_LARGE",
              "RATE_LIMITED",
              "UPSTREAM_UNAVAILABLE",
              "INTERNAL",
              "NETWORK_ERROR",
              "TIMEOUT",
              "ABORTED",
              "UNKNOWN"
            ],
            "example": "NOT_FOUND",
            "description": "Stable machine-readable code — branch on this, never on message"
          },
          "message": {
            "type": "string",
            "example": "Document 3f2a… not found"
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "description": "Machine-readable extras: validation errors[], conflict currentHeadRevisionId/comparisonUrl, …"
          },
          "path": {
            "type": "string",
            "example": "/v1/documents/3f2a"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-02T12:00:00.000Z"
          },
          "requestId": {
            "type": "string",
            "description": "Correlation id — echoed from the x-request-id request header (also set as a response header)"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "path",
          "timestamp",
          "requestId"
        ]
      }
    }
  }
}